Fix test code to eliminate warnings from -Wall -Wextra

+++ test/ffs.c
@@ -43,7 +43,7 @@ long long int
 rand_long_long(void)
 {
 	unsigned long long int r = 0;
-	int x;
+	size_t x;
 
 	for (x = 0; x < sizeof(r) / 4; x++)
 		r = (r << 32) | (lrand48() & 0xffffffffL);
@@ -98,7 +98,7 @@ check(long long int x, int got, int expect, char *which)
 int
 main(void)
 {
-	int i;
+	size_t i;
 	int ret = 0;
 
 	/* Test zero */
